requirements.txt
# ═══════════════════════════════════════════════════════════════
# BioInsight AI - Python Dependencies
# ═══════════════════════════════════════════════════════════════

# PDF Processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

# Text Processing & NLP
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1

# Embeddings - PubMedBERT
sentence-transformers>=2.2.0
torch>=2.0.0

# Vector Database
chromadb>=0.4.0

# LLM - Google Gemini
google-generativeai>=0.3.0
langchain-google-genai>=0.0.5

# LLM - Anthropic Claude
anthropic>=0.75.0
langchain-anthropic>=1.3.0

# ═══════════════════════════════════════════════════════════════
# Data Science / ML
# ═══════════════════════════════════════════════════════════════
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# RNA-seq Analysis
statsmodels>=0.14.0
networkx>=3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# ═══════════════════════════════════════════════════════════════
# Web API & HTTP
# ═══════════════════════════════════════════════════════════════
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
aiohttp>=3.9.0
httpx>=0.26.0
requests>=2.31.0

# ═══════════════════════════════════════════════════════════════
# Bio Tools
# ═══════════════════════════════════════════════════════════════
biopython>=1.81

# ═══════════════════════════════════════════════════════════════
# Utilities
# ═══════════════════════════════════════════════════════════════
python-dotenv>=1.0.0
tqdm>=4.66.0
pydantic>=2.0.0

# ═══════════════════════════════════════════════════════════════
# Development & Testing
# ═══════════════════════════════════════════════════════════════
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
